Datatype of various backend generated ids was changed from Guid to string. The SharesRepository is implemented. Errors in the context class arising from the previous commit is handled here. A change in the datatype of the sharecount from int  to decimal in the sql commands

+++ SACCO System/SQL/sqlCommands.sql
@@ -125,7 +125,7 @@ CREATE TABLE Dividend (
 CREATE TABLE Shareholder (
     Shareholder_ID CHAR(36) PRIMARY KEY,
     Member_ID INT,
-    Share_Count INT,
+    Share_Count DECIMAL,
     FOREIGN KEY (Member_ID) REFERENCES Member(Member_ID)
 );
